+++ src/chains/Arbitrum/Common.sol
@@ -11,6 +11,7 @@ import {DodoV2, IDodoV2} from "../../core/DodoV2.sol";
 import {UniswapV4} from "../../core/UniswapV4.sol";
 import {IPoolManager} from "../../core/UniswapV4Types.sol";
 import {BalancerV3} from "../../core/BalancerV3.sol";
+import {NativeV2} from "../../core/NativeV2.sol";
 import {FreeMemory} from "../../utils/FreeMemory.sol";
 
 import {ISettlerActions} from "../../ISettlerActions.sol";
@@ -53,7 +54,8 @@ abstract contract ArbitrumMixin is
     DodoV1,
     DodoV2,
     UniswapV4,
-    BalancerV3
+    BalancerV3,
+    NativeV2
 {
     constructor() {
         assert(block.chainid == 42161 || block.chainid == 31337);
@@ -115,6 +117,10 @@ abstract contract ArbitrumMixin is
                 abi.decode(data, (IERC20, uint256, IDodoV1, bool, uint256));
 
             sellToDodoV1(sellToken, bps, dodo, quoteForBase, minBuyAmount);
+        } else if (action == uint32(ISettlerActions.NATIVEV2.selector)) {
+            (address router, uint256 bps, bytes memory tradeData) = abi.decode(data, (address, uint256, bytes));
+
+            sellToNativeV2(router, bps, tradeData);
         } else {
             return false;
         }